S01E03/requirements.txt
requests
python-dotenv
loguru
openai